refactor and name change
