+++ cmd/horcrux/cmd/start.go
@@ -4,13 +4,12 @@ import (
 	"fmt"
 	"os"
 
+	cconfig "github.com/strangelove-ventures/horcrux/src/config"
 	"github.com/strangelove-ventures/horcrux/src/connector"
 
 	cometlog "github.com/cometbft/cometbft/libs/log"
 	"github.com/cometbft/cometbft/libs/service"
 	"github.com/spf13/cobra"
-	cconfig "github.com/strangelove-ventures/horcrux/src/config"
-	"github.com/strangelove-ventures/horcrux/src/node"
 )
 
 func startCmd() *cobra.Command {
@@ -23,7 +22,7 @@ func startCmd() *cobra.Command {
 			out := cmd.OutOrStdout()
 			logger := cometlog.NewTMLogger(cometlog.NewSyncWriter(out))
 
-			err := node.RequireNotRunning(logger, config.PidFile)
+			err := RequireNotRunning(logger, config.PidFile)
 			if err != nil {
 				return err
 			}
@@ -79,7 +78,7 @@ func startCmd() *cobra.Command {
 				return fmt.Errorf("failed to start remote signer(s): %w", err)
 			}
 
-			node.WaitAndTerminate(logger, services, config.PidFile)
+			WaitAndTerminate(logger, services, config.PidFile)
 
 			return nil
 		},
