--- src/cosigner/icosigner.go
@@ -1,1 +1,0 @@
-package cosigner
